[Algorithm] Price without tesseract

+++ yourbudget/receipt_back/tasks.py
@@ -11,18 +11,20 @@ import os
 
 
 @app.task
-def save_trip(username, img_path):
+def save_trip(username, img_path, save=True):
     logging.info('saving trip for {0} from {1}'.format(username, img_path))
 
     shopping_trip = ReceiptReader.convert_to_receipt(img_path)
 
     logging.info('finished parsing')
+    logging.info('i got ' + shopping_trip.to_json())
 
-    from mongoengine import connect
-    connect('myNewDatabase')
+    if save == 'True':
+        from mongoengine import connect
+        connect('myNewDatabase')
 
-    history = UserData.get_history(username)
-    history.all_trips.append(shopping_trip)
-    history.save()
+        history = UserData.get_history(username)
+        history.all_trips.append(shopping_trip)
+        history.save()
 
-    os.remove(img_path)
+        os.remove(img_path)
